Local sources of sphinxgallery 0.0.6
