+++ examples/README.txt
@@ -1,3 +1,14 @@
+.. warning::
+
+    If you want to run the examples you need to make sure you have
+    installed nilearn by following the instructions :ref:`there
+    <installation>`.
+
+.. note::
+
+    A few examples may not run with scikit-learn versions older than
+    0.14.1.
+
 General examples
 ----------------
